fix: move middleware into different folder
